requirements.txt
fastapi==0.111.0 # Or whatever latest stable is
uvicorn==0.30.1 # Or whatever latest stable is
python-multipart==0.0.9 # Or whatever latest stable is
google-generativeai==0.7.1 # Or whatever latest stable is
python-dotenv==1.0.1 # Or whatever latest stable is
requests==2.32.3 # Or whatever latest stable is
